2nd draft: accountid, number_format(), spaces in filters, no pagination, account balance report
Add the account id to any page/report that has the account name.  including
being able to select an account by id on the add action page.
Use number_format() for amounts.  Especially the total amount on the
transactions report, where math and weird precision display can happen.
Fix spaces in filters on tables on the account, action, and contact pages.
Remove pagination from the receipts and 500+ reports.
Add an empty location option to the 500+ report.
Have default values for the dates on the transaction report.
Add the school account balance report.